rpm/requirements.txt
pyyaml
jinja2
bs4

